CSP-1419 Moved get AccountProvider to write repository
