+++ src/SFA.DAS.PR.Application.UnitTests/Permissions/Commands/PostPermissions/PostPermissionsCommandHandlerTests/PostPermissionsCommandHandlerIntegrationTests.cs
@@ -250,7 +250,6 @@ public class PostPermissionsCommandHandlerIntegrationTests
             permissionsWriteRepository,
             permissionsAuditWriteRepository,
             context,
-            Mock.Of<IAccountProviderReadRepository>(),
             Mock.Of<IMessageSession>()
         );
     }
