+++ src/SFA.DAS.PR.Domain/Interfaces/IAccountProviderWriteRepository.cs
@@ -5,4 +5,5 @@ namespace SFA.DAS.PR.Domain.Interfaces;
 public interface IAccountProviderWriteRepository
 {
     Task<AccountProvider> CreateAccountProvider(long ukprn, long accountId, CancellationToken cancellationToken);
+    Task<AccountProvider?> GetAccountProvider(long? ukprn, long accountId, CancellationToken cancellationToken);
 }
